Moved yim tests to yim folder
