+++ app/src/test/java/org/listenbrainz/android/yim/YimViewModelTest.kt
@@ -1,4 +1,4 @@
-package org.listenbrainz.android
+package org.listenbrainz.android.yim
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -11,14 +11,14 @@ import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
-import org.listenbrainz.sharedtest.mocks.MockAppPreferences
-import org.listenbrainz.sharedtest.mocks.MockYimRepository
 import org.listenbrainz.android.util.Constants.Strings.STATUS_LOGGED_IN
 import org.listenbrainz.android.util.Resource
 import org.listenbrainz.android.viewmodel.YimViewModel
+import org.listenbrainz.sharedtest.mocks.MockAppPreferences
+import org.listenbrainz.sharedtest.mocks.MockYimRepository
+import org.listenbrainz.sharedtest.testdata.YimRepositoryTestData.testYimData
 import org.listenbrainz.sharedtest.utils.AssertionUtils.checkYimAssertions
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testUsername
-import org.listenbrainz.sharedtest.testdata.YimRepositoryTestData.testYimData
 
 class YimViewModelTest{
